Allow model objects to be passed instead of an initialized model.
New test for the pagination plugin helped uncover a shared object among multiple instances. When multiple instances of the same view are used in plugins with different configuration, a model object needs to be configured so that each instance has its own unique model instance.

Other views continue to behave as expected.